mcp-app-widget-capture: extract from evals-runner into shared helper
The captureEvalTraceWidgetSnapshots function in evals-runner.ts was
file-private but the operation itself is generic: walk an AI SDK
transcript for MCP App tool calls, fetch the widget HTML via
MCPClientManager.readResource(), upload it to Convex storage, and
return EvalTraceWidgetSnapshot[]. Nothing about it is eval-specific.

Pull it (plus its private helpers: collectToolSnapshotSources,
normalizeWidgetCsp, normalizeWidgetPermissions,
extractHtmlFromResourceContent, uploadWidgetHtmlBlob, and the small
ToolSnapshotSource walker utilities) into a new
server/utils/mcp-app-widget-capture.ts and re-export as
captureMcpAppWidgetSnapshots. Log tag changes from [evals] to
[mcp-app-widget-capture] so non-eval callers don't look misattributed.

evals-runner.ts drops the moved code, dependent imports
(getToolUiResourceUri, isMcpAppTool, injectOpenAICompat), the
now-orphaned isRecord/readRecordString helpers, and renames its six
internal call sites. Pure code move: no behavior change for evals.

The synthetic-session runner (next commit) consumes the shared helper
so chatbox synthetic sessions can persist widget snapshots on the same
path the eval runner already uses.
